requirements.txt
streamlit
langchain==0.1.20
langchain-openai
langchain-community
chromadb
pypdf
python-dotenv
langchain-text-splitters